change sessionid to 16-bit

+++ plugin-echoprobe/EchoProbeRequestSegment.hxx
@@ -20,7 +20,7 @@ enum class EchoProbeReplyStrategy: uint8_t {
 struct EchoProbeRequest {
     bool deviceProbingStage = false;
     EchoProbeReplyStrategy replyStrategy = EchoProbeReplyStrategy::ReplyWithCSI;
-    uint32_t sessionId = 0;
+    uint16_t sessionId = 0;
     int16_t repeat = 1;
     int16_t ackMCS = -1;         // 0 to11 are OK, negative means use default (maybe mcs 0).
     int16_t ackNumSTS = -1;
